+++ app/src/main/java/free/rm/skytube/gui/activities/ShareBookmarkActivity.java
@@ -1,16 +1,24 @@
+/*
+ * SkyTube
+ * Copyright (C) 2020  Zsombor Gegesy
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation (version 3 of the License).
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package free.rm.skytube.gui.activities;
 
-import android.content.Intent;
-import android.os.Bundle;
-import android.widget.Toast;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-import org.schabi.newpipe.extractor.StreamingService;
+import androidx.annotation.StringRes;
 
 import free.rm.skytube.R;
-import free.rm.skytube.app.SkyTubeApp;
 import free.rm.skytube.businessobjects.YouTube.YouTubeTasks;
 import free.rm.skytube.businessobjects.YouTube.newpipe.ContentId;
 import io.reactivex.rxjava3.disposables.CompositeDisposable;
@@ -19,31 +27,20 @@ import io.reactivex.rxjava3.disposables.CompositeDisposable;
  * Activity that receives an intent from other apps in order to bookmark a video from another app.
  * This Activity uses a transparent theme, and finishes right away, so as not to take focus from the sharing app.s
  */
-public class ShareBookmarkActivity extends AppCompatActivity {
+public class ShareBookmarkActivity extends ExternalUrlActivity {
     private final CompositeDisposable compositeDisposable = new CompositeDisposable();
 
     @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        if(getIntent() != null) {
-            String textData = getIntent().getStringExtra(Intent.EXTRA_TEXT);
-            ContentId contentId = SkyTubeApp.parseUrl(this, textData, false);
-            if (contentId.getType() == StreamingService.LinkType.STREAM) {
-                compositeDisposable.add(YouTubeTasks.getVideoDetails(this, contentId)
-                        .subscribe(video -> {
-                            if (video != null) {
-                                video.bookmarkVideo(ShareBookmarkActivity.this);
-                            } else {
-                                invalidUrlError();
-                            }
-                            finish();
-                        }));
-            } else {
-                SkyTubeApp.openUrl(this, textData, false);
-                finish();
-            }
-        }
+    void handleVideoContent(ContentId contentId) {
+        compositeDisposable.add(YouTubeTasks.getVideoDetails(this, contentId)
+                .subscribe(video -> {
+                    if (video != null) {
+                        video.bookmarkVideo(free.rm.skytube.gui.activities.ShareBookmarkActivity.this);
+                    } else {
+                        invalidUrlError();
+                    }
+                    finish();
+                }));
     }
 
     @Override
@@ -52,8 +49,8 @@ public class ShareBookmarkActivity extends AppCompatActivity {
         super.onDestroy();
     }
 
-    private void invalidUrlError() {
-        Toast.makeText(this, R.string.bookmark_share_invalid_url, Toast.LENGTH_LONG).show();
-        finish();
+    @StringRes
+    int invalidUrlErrorMessage() {
+        return R.string.bookmark_share_invalid_url;
     }
 }
